backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv

# fastapi[all]